Исправь splash на intro
