+++ js/application.js
@@ -5,19 +5,15 @@ import HeaderController from './header-controller.js';
 import GameModel from './game-model.js';
 import GameController from './game-controller.js';
 import StatsController from './stats-controller.js';
-import SplashController from './splash-controller.js';
 
 class Application {
-  static start() {
-    SplashController.showSplash();
-  }
 
   static renderIntro() {
     IntroController.showIntro(Application.renderGreeting);
   }
 
   static renderGameCB(name) {
-    const newGame = new GameController(new GameModel(name, SplashController.getQuestions()));
+    const newGame = new GameController(new GameModel(name, IntroController.getQuestions()));
     newGame.renderGameState(Application.renderGreeting, Application.renderStatsCB);
   }
