+++ js/intro-view.js
@@ -3,6 +3,8 @@ import AbstractView from './abstract-view.js';
 class ViewIntro extends AbstractView {
   constructor(cb) {
     super();
+    // this.cursor = 0;
+    // this.symbolsSeq = `/—\\|`;
     this.tag = `section`;
     this.classList = [`intro`];
     this.cb = cb;
@@ -14,6 +16,16 @@ class ViewIntro extends AbstractView {
       <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>`;
 
     return intro;
+  }
+
+  start() {
+    // this.cursor = ++this.cursor >= this.symbolsSeq.length ? 0 : this.cursor;
+    // this.element.textContent = this.symbolsSeq[this.cursor];
+    this.timeout = setTimeout(() => this.start(), 50);
+  }
+
+  stop() {
+    clearTimeout(this.timeout);
   }
 
   bind() {
